changed login and register page to be components and added a AuthView page
