--- dev/client/src/views/RegisterView.vue
@@ -1,105 +1,0 @@
-<script setup lang="ts">
-import { defineComponent } from "vue";
-import DefaultButton from "../components/DefaultButton.vue";
-</script>
-<script lang="ts">
-export default defineComponent({
-  data() {
-    return {
-      user: {
-        email: "",
-        password: "",
-      },
-    };
-  },
-  methods: {
-    async loginUser() {
-      const myHeaders = new Headers();
-      myHeaders.append("Content-Type", "application/json");
-
-      const requestOptions = {
-        method: "POST",
-        headers: myHeaders,
-        body: JSON.stringify(this.user),
-      };
-
-      const response = await fetch(
-        "https://dont-pani.cc/api/session",
-        requestOptions
-      );
-
-      if (response.status === 201) {
-        console.log("Successfully logged in");
-        this.user = { email: "", password: "" }; // Clear user form data
-      } else {
-        console.log("Failed to login");
-      }
-    },
-    async registerUser() {
-      console.log(this.user);
-      const myHeaders = new Headers();
-      myHeaders.append("Content-Type", "application/json");
-
-      const requestOptions = {
-        method: "POST",
-        headers: myHeaders,
-        body: JSON.stringify(this.user),
-      };
-
-      const response = await fetch(
-        "https://dont-pani.cc/api/user",
-        requestOptions
-      );
-
-      if (response.status === 201) {
-        console.log("Successfully registered");
-        await this.loginUser();
-      } else {
-        console.log("Failed to register");
-      }
-    },
-  },
-  computed: {},
-});
-</script>
-<template>
-  <div class="login-page">
-    <h1>Register</h1>
-    <div>
-      <input placeholder="Email" v-model="user.email" />
-    </div>
-    <div>
-      <input placeholder="Password" v-model="user.password" />
-    </div>
-    <DefaultButton msg="Register" @click="registerUser()" />
-    <!-- <button>Log In</button> -->
-    <p>or</p>
-    <RouterLink to="/Login"><DefaultButton msg="Login" /></RouterLink>
-  </div>
-</template>
-
-<style scoped>
-.login-page {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 400px;
-  height: 400px;
-  gap: 1em;
-  font-size: 20px;
-}
-.login-page div input {
-  width: 200px;
-  height: 40px;
-  border-radius: 5px 5px 5px 5px;
-  background-color: var(--vt-c-white-mute);
-}
-button {
-  width: 200px;
-  padding: 0.4rem;
-}
-/* #newBtn{
-  width: 200px;
-} */
-</style>
